Ignore application.properties et retire du suivi

+++ src/main/java/ma/osbt/entitie/Reservation.java
@@ -21,7 +21,7 @@ public class Reservation  {
 	 @Id
 	 @GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
-	private int prix;
+	private Double prix;
 	private String statut;
 	private Date dateReservation;
 	@ManyToOne
